Add functionality to allow setting a query string in a GET request
Up until this point there was no need for the REST API to receive query parameters in a GET request. However, the Lookup endpoint does have query parameters in the GET request, so something needed to be changed in the RestClient and in the Resource classes.

Note that the implementation of the QueryString property is really basic, as it just returns an empty string. The ideal way this would have been set up, if we would have checked all the properties of a given object, and just turn them into a query string automatically. However, given that we use JsonPropertyAttribute and the EnumMemberAttribute for Enums getting the actual values is harder.

Another issue is that we don't necessarily need all the properties of an object, only some.

The solution is to have individual resources send their own query string parameters. An example of this can be found in the Lookup resource in the next commit.

Here's some inspiration on how this could be solved more automagically in the future: https://ole.michelsen.dk/blog/serialize-object-into-a-query-string-with-reflection.html